"Path browser" - 4 scrolled lists displaying:     directories on sys.path     modules in selected directory     classes in selected module     methods of selected class
Sinlge clicking in a directory, module or class item updates the next
column with info about the selected item.  Double clicking in a
module, class or method item opens the file (and selects the clicked
item if it is a class or method).

I guess eventually I should be using a tree widget for this, but the
ones I've seen don't work well enough, so for now I use the old
Smalltalk or NeXT style multi-column hierarchical browser.