Bring PEP 453 What's New entry up to date
We changed a few details of the pip integration over the course
of the Python 3.4 beta and release candidate cycle, but I hadn't
updated the What's New entry appropriately.

The Py2/Py3 split on POSIX and the generally challenges of
dealing with parallel installs means we haven't attained
complete consistency across all environments, but hopefully
this updated text is clearer about the end result.

Relevant to issue #20909.