Implement a basic PyUnicode_FSConverter for Tauthon
For ease of backporting, it is useful to have a PyUnicode_FSConverter.

Bringing in the entire PyUnicode_FS* logic from Python3 was judged to be
both much larger than necessary for my use (POSIX *at() function
backports) and ran the risk of introducing encoding behavior
changes from Py2.

This implementation strives to be compatible with existing encoding
treatment in Tauthon's posixmodule (i.e., "et" with
Py_FileSystemDefaultEncoding).  It is derived from the Python3 function
of the same name, but adapted to follow the same logic as Tauthon "et"
conversions to Py_FileSystemDefaultEncoding.  Notably, unlike Py3.1+ it
does not round-trip the WTF-8 subset known as UTF-8b.  I don't think
there is any reason Tauthon could not do this eventually, but when it is
done it should be done consistently.

Handling of os.PathLike inputs via PyOS_FSPath() (originally from Py3.6)
seems nice to have, but was left unimplemented for now.