setup.py: Small metadata update, replace official Python links
Make the package URL point to Tauthon's GitHub repository, and remove the
official Python's dev email. Add a small paragraph on Tauthon in the summary.

Did not change the project name to Tauthon, since doing this causes the final
egg to be renamed, and has the side-effect of making pip list Tauthon as an
installed package, which it doesn't do for "Python" (it is probably testing the
name, this should be patched for maximum compatibility when changing it).