bpo-37820: urllib: Don't "unwrap" passed URLs
Unwrap means removing enclosing ('<','>'), if any, and then a possible 'URL:'
prefix. The 'unwrap()' function (actually, its code, which was originally
inlined) comes from Python's very beginnings, and the only normative document I
could relate it to is RFC3986, Appendix C ("Delimiting a URI in Context"). It
states that these additions can be used to delimit URIs inside other
documents (and incidentally specifies that adding 'URL:' is not recommended).

In any case, we expect users to pass true URLs here, and not "delimited" ones,
so just don't pre-process passed URLs like this.

This closes the possibility to open local files via such "URLs" (as just
explained, they are in fact not):
<local_file>
URL:local_file
<URL:local_file>