bpo-43223: Replace '//' at start of path to avoid redirection to another host
Inspired by the patch proposed by Hamza Avvan and amended by Victor Stinner in
the aforementioned bug.

Upstream bug is still open. A glance at RFC 3986 makes me think that a more
proper fix would be simply to forbid such URLs (they appear not to be valid
ones indeed), and to implement all checks in urlparse/urlsplit instead. But I
have no time for this, and this change seems good enough. We'll see what
upstream does in the end.