[2.8] bpo-42938: Replace snprintf with Python unicode formatting in ctypes param reprs. (24239)
(cherry picked from commit 916610ef90a0d0761f08747f7b0905541f0977c7)

Fixed the new test.

While here, imported small changes in the representation for better debugging
and to ease future porting. Ported 'is_literal_char' in particular, in a
straightforward way.

Build on Windows: Seems that the compiler used in the CI is configured to be
strict about declarations in the middle of blocks, so avoid that.

Co-authored-by: Benjamin Peterson <benjamin@python.org>